no-floating-promises is now recommended.
This does now apply to client-side configs as well which is new. There are some
places in client-side code where we had floating promises we didn't want so I
think enabling here is good. Additionally the void keyword can be used to
explicitly indicate you want to don't want to wait for the promise.
i.e. `void asyncFn();`